+++ RELEASE-NOTES.md
@@ -17,6 +17,7 @@ Latest release:
 * Dropped class aliases deprecated since Diff 1.0
 * Removed `ListDiff` and `MapDiff`, deprecated since Diff 0.5
 * Removed `ListDiffer::MODE_NATIVE` and `ListDiffer::MODE_STRICT`, deprecated since Diff 0.8
+* Removed `MapDiffer::setComparisonCallback` in favour of a new constructor argument
 
 ## Version 2.1 (2016-09-01)